Revert "Bug 1998457, Bug 1998451, Bug 1998452, Bug 1998161 - Part 6: Add codegen tests for division/remainder with constants. r=spidermonkey-reviewers,iain" for causing SM failures on wasm.js
This reverts commit 6c4047cbb453260868cf6879dbd50cbf0bb871eb.

Revert "Bug 1998457 - Part 5: Don't generate unnecessary code for modulus by one. r=spidermonkey-reviewers,iain"

This reverts commit d48bfd57f35248fecc074c5cb20080ccf1085d65.

Revert "Bug 1998457 - Part 4: Optimise int64 divison and modulus with constants in arm64. r=spidermonkey-reviewers,iain"

This reverts commit 7768107fde9612324a6d1300c3592c1389198264.

Revert "Bug 1998457 - Part 3: Optimise int64 divison and modulus with constants in x64. r=spidermonkey-reviewers,iain"

This reverts commit 09e6a576a304ef061aac062e67b5b776653440b6.

Revert "Bug 1998457 - Part 2: Add imulq and umulq. r=spidermonkey-reviewers,iain"

This reverts commit 81a2dd4de1abc2b3fdfb87a2bb509cd7e32142ed.

Revert "Bug 1998457 - Part 1: Extend ReciprocalMulConstants for Int64 division. r=spidermonkey-reviewers,iain"

This reverts commit 34de8212a70815d504d03d80ef1b411e0bb9f881.

Revert "Bug 1998451: Improve performance of codegen tests. r=bvisness"

This reverts commit 1e220e42dc45ebb2161bba9fd8fda308e64ee3e4.

Revert "Bug 1998452: Avoid REX prefix for And with unsigned 32-bit immediate. r=spidermonkey-reviewers,iain"

This reverts commit 8e2181952db0848c3c58b2da236ffe2c772cc632.

Revert "Bug 1998161 - Part 4: Merge shifts for division by constants for x64. r=iain"

This reverts commit 7559329c7d447caf529e9b3745d40d53668e5b9b.

Revert "Bug 1998161 - Part 3: Remove register constraints for div/mod with constants on x64. r=spidermonkey-reviewers,iain"

This reverts commit de336ac8ca77b5fe259109d21772f25fca44d48d.

Revert "Bug 1998161 - Part 2: Use register allocator to place lhs operand for div/mod into eax. r=spidermonkey-reviewers,iain"

This reverts commit bc54bc3927d782af08d4f84a73dc417d99b52343.

Revert "Bug 1998161 - Part 1: Port arm64 changes to x86/x64. r=spidermonkey-reviewers,iain"

This reverts commit 92f6d9ca04e5c9e48c3c9b1eed3b2bdb2615be72.

UltraBlame original commit: b8da64e1b7acced081d17251b8707e8fc9803150

+++ js/src/jit/MacroAssembler.cpp
@@ -2062,8 +2062,7 @@ void MacroAssembler::loadInt32ToStringWithBase(
       
       
 
-      auto rmc = ReciprocalMulConstants::computeUnsignedDivisionConstants(
-          uint32_t(base));
+      auto rmc = ReciprocalMulConstants::computeUnsignedDivisionConstants(base);
 
       
       mulHighUnsigned32(Imm32(rmc.multiplier), input, scratch1);
